+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Fixed bug in GPU training. (#22)
 * Added GPU CI. (#22)
+* Explicit error in case of missing values (@cregouby, #24)
 
 # tabnet 0.1.0